--- Assets/PulseSequencer/Editor/AudioFilters/VolumeEnvelopeInspector.cs
@@ -1,10 +1,0 @@
-﻿using UnityEngine;
-using UnityEditor;
-using System.Collections;
-
-namespace DerelictComputer
-{
-	public class VolumeEnvelopeInspector : Editor 
-	{
-	}
-}
