--- Assets/PulseSequencer/Editor/AudioFilters/VolumeEnvelopeInspector.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 3043738728b524f49a0394a490950515
-timeCreated: 1439334582
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
