--- Assets/PulseSequencer/Editor/Core/SamplerInspector.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 57461b29eeb524a5a9438e87f34d0c1d
-timeCreated: 1439302791
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
