--- Assets/PulseSequencer/Scripts/PatternFollowers.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: b06d6edfbffe47740bb796e0fe1becec
-folderAsset: yes
-timeCreated: 1439304104
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
