+++ Assets/PulseSequencer/Scripts/VisualEffects.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: fb289d1da33514981aecd06c8731191e
+guid: 1a738c68ce32e465c8fa3d3a2e537c96
 folderAsset: yes
-timeCreated: 1439334569
+timeCreated: 1439422117
 licenseType: Pro
 DefaultImporter:
   userData:
