+++ Assets/PulseSequencer/Scripts/VisualEffects/AnimationTriggerFollower.cs
