+++ Assets/PulseSequencer/Scripts/VisualEffects/LightLerpEffect.cs
